Update todo_list_item.tsx
I actually think this interior item component is too small to bother putting into another component, but I do want the practice of moving things into these other components.

So I took out the checkmark functionality and the only thing that the Item component controls is the editability and the ItemName. Then I put this next to the currrent to-do list item controlled by ternary statement.

Moving this into another component like this would be better for something like a Trello clone that has an "interior item view" or something similar.